feat: 알림 권한 기능 구현

+++ bangtong/back/src/main/java/com/jisang/bangtong/dto/user/ProfileDto.java
@@ -4,7 +4,8 @@ import lombok.Data;
 
 @Data
 public class ProfileDto {
-    private Integer userId;
+    private Long userId;
     private String mediaPath;
     private String userNickname;
+
 }
